Init Commit!! :dromedary_camel:
The goal of this project is to create a calculator implementation using
fun OOP approaches, like a Command Pattern. I want to be able to store
Commands as well and perhaps undo them, just having fun with it.

Perhaps I can use Swing to make a GUI as well, but that might be a
headache.