Baby's first JUnit Tests!!!!
This is so exciting! Now I feel like this redefines my idea of
test-driven development. I'm not sure how deep the tests should be, or
if I should be declaring multiple instances of the test class object in
the same test method... That's stuff to learn though.

My next goal is going to be to write tests for the rest of my classes.
Again, I'm smiling. This is so cool, writing, testing (should swap that
last one with this one), and ultimately pushing code is pretty awesome
so far. What a workflow.